requirements.txt
xai-sdk==1.4.0
python-dotenv==1.2.1

